Conflicts:
	IrrigationAdvisor/Controllers/Reports/ReportPivotStateController.cs
	IrrigationAdvisor/Views/ReportPivotState/ReportPivotState.cshtml

+++ IrrigationAdvisor.WebApi/Models/AdviceViewModel.cs
@@ -5,7 +5,7 @@ namespace IrrigationAdvisor.WebApi.Models
     public class AdviceViewModel
     {
         public string IrrigationType { get; set; }
-        public decimal Quantity { get; set; }
+        public double Quantity { get; set; }
         public DateTime Date { get; set; }
     }
 }
